// app/interfaces/MainComponent/ContactOfficecontactcomponent.ts
export interface IContactUscontact {
    _id: string;
    id?: string;
    title?: string;
    image?: string;
    description?: string;
    email?: string;
    number?: string;
}



export interface IContactUscontactInput {
    id?: string;
    title?: string;
    image?: string;
    description?: string;
    email?: string;
    number?: string;
}